requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
iso8601==2.1.0
pydantic==2.11.5
pydantic_core==2.33.2
pypika-tortoise==0.6.1
pytz==2025.2
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.5
pydantic_core==2.33.2
pypika-tortoise==0.6.1
pytz==2025.2
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
